refactor: MailMandalartGrid, MandalartReminder는 application 계층으로 이동
